Moves non-server files into lib/ fixes like to index - can serve now.
Created the rest of the backend project structure with a lib, public,
and test folders. Fixed a path to the index page (up a dir) so running
node server will now server back a flat html page (links are all off).

Backend is done enough to start porting the front-end to React. Yay!