Adds an About component which holds the bio components
The about is a fairly dumb component, since the data on it can't
really change. The bio info is stored elsewhere as a constant, and
is pulled in by the bio instead of being passed by About (which
really has no reason to know about the bios).

Might be possible to do this with composition using props.children,
but sticking to this model for now.