Creates a dropdown menu for the county using a prop location
Need to have a way to manage which side of the app is changing data,
and to have the text match the current versus the destination side.
This tries to have a textual separation based on whether the component
is passed the prop location "current" or "destination".

Not sure if this is scalable for the app, but it works for the text
change issue atm.