Pulls other curr/dest select components together -- css not working
The three dropdowns (state, county, and city) are all loading with
their css, and are getting data passed in through props.

The three data display components (IncomeDisplay, HousePrices, and
RentalMarket) are all displaying, minus the data that should be
being passed to them on props.

The css for the left/right side (flex containers, 50% width, color)
isn't working yet - not clear where to put the styles to get them
to work again. Debugging.