Deletes component-ified code from old index.html
I made a copy of the old index.html page so that I could cut big
chunks of it out and convert them into components (reusing the html
and styles where possible). This made it a lot quicker to get the
repetitive bits or former Handlebars code converted to React
components. Still a bit left to convert.