Adds a Navbar component which uses React-router to link/load components
Package.json now include react-router and react-router-dom to start using
the router functionality for linking.

The BrowserRouter parent tag is in the index.js to wrap the entire project
in the router.

The Navbar component changed the old static a href links into router <Link>
which use a second <Route> tag to load a component. Loading the About and
the ReactTransplantr components works - but there's a visual effect mistake
with a horizontal green bar below the hero banner. Debugging later.