Moves Routes from Nav to index, fixes bug in Links so links show up
The Navbar links weren't showing up because I replaced the <li> tags
with the <Link> tags without thinking about the fact that the styling
in the <ul> was on the <li> so not only did it lose styles, the id
and  class were on the wrong tag also breaking the functionality of
the links. Re-adding the li and putting the attributes on it fixes it.

I wanted to keep the routes in the top level file so it would be easier
to see what components were being loaded -- having that with the links
doesn't make sense when trying to tease out the functionality of the
app.