Adds the census values for states, and puts all counties into array
The old app was using a call to the census to load the data for the
dropdowns for counties, which adds an unexpected delay. This moves
all of the data from Wikipedia into a const file (with a script to
clean it all up). Now the counties can load when the state changes.