Uses state selection to fill county dropdown
Uses the e.target.value state code to look up the state name from
the STATES, and then used the state name to lookup the counties
from the county dict, then updates the state for the selected state
code, state name, and counties array.

The dropdown weren't updating their values correctly because I
needed to pass the value from state to the selects via the props.
Setting the value from the state makes it update correctly (rather
than with a lag).